refactor: moved sync run table item to new component
